Set up skeleton for media abstraction.
