+++ src/media/AudioDisplayComponent.h
@@ -1,0 +1,42 @@
+#pragma once
+
+#include <juce_audio_utils/juce_audio_utils.h>
+
+#include "MediaDisplayComponent.h"
+
+
+class AudioDisplayComponent : public MediaDisplayComponent
+{
+public:
+    ~AudioDisplayComponent()
+    {
+        // TODO
+    }
+
+    void setupDisplay()
+    {
+        // TODO
+    }
+
+    std::string getSupportedExtensions()
+    {
+        // TODO
+    }
+
+    void loadMediaFile(const URL& filePath)
+    {
+        // TODO
+    }
+
+    void startPlaying()
+    {
+        // TODO
+    }
+
+    void stopPlaying()
+    {
+        // TODO
+    }
+
+private:
+};
