+++ src/media/MidiDisplayComponent.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#include "pianoRoll/PianoRollEditorComponent.hpp"
+#include "MediaDisplayComponent.h"
+
+
+class MidiDisplayComponent : public MediaDisplayComponent
+{
+public:
+    ~MidiDisplayComponent()
+    {
+        // TODO
+    }
+
+    void setupDisplay()
+    {
+        // TODO
+    }
+
+    std::string getSupportedExtensions()
+    {
+        // TODO
+    }
+
+    void loadMediaFile(const URL& filePath)
+    {
+        // TODO
+    }
+
+    void startPlaying()
+    {
+        // TODO
+    }
+
+    void stopPlaying()
+    {
+        // TODO
+    }
+
+private:
+};
